+++ monkeytype/tracing.py
@@ -190,9 +190,9 @@ class CallTracer:
     def __init__(
         self,
         logger: CallTraceLogger,
+        max_typed_dict_size: int,
         code_filter: Optional[CodeFilter] = None,
         sample_rate: Optional[int] = None,
-        max_typed_dict_size: Optional[int] = None,
     ) -> None:
         self.logger = logger
         self.traces: Dict[FrameType, CallTrace] = {}
@@ -269,13 +269,13 @@ class CallTracer:
 @contextmanager
 def trace_calls(
     logger: CallTraceLogger,
+    max_typed_dict_size: int,
     code_filter: Optional[CodeFilter] = None,
     sample_rate: Optional[int] = None,
-    max_typed_dict_size: Optional[int] = None,
 ) -> Iterator[None]:
     """Enable call tracing for a block of code"""
     old_trace = sys.getprofile()
-    sys.setprofile(CallTracer(logger, code_filter, sample_rate, max_typed_dict_size))
+    sys.setprofile(CallTracer(logger, max_typed_dict_size, code_filter, sample_rate))
     try:
         yield
     finally: